pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-edge-ai"
version = "0.1.0"
description = "IoT Edge Device Simulator with MQTT Telemetry"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache License 2.0"}
authors = [
    {name = "UFBA - LLM Edge AI Team"}
]
keywords = ["iot", "edge-computing", "mqtt", "simulation", "docker"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "paho-mqtt==1.6.1",
    "pandas==1.5.3",
    "python-dotenv==1.0.0",
    "numpy==1.24.3",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/alanssoares/llm-edge-ai"
Repository = "https://github.com/alanssoares/llm-edge-ai"

[project.scripts]
generate-compose = "src.generate_compose:main"
device-simulator = "src.device_simulator:main"
mqtt-consumer = "src.mqtt_consumer:main"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[[tool.mypy.overrides]]
module = "paho.mqtt.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pandas.*"
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
